+++ dio_cache_interceptor/lib/src/model/nullable.dart
@@ -1,0 +1,9 @@
+class Nullable<T> {
+  final T? _value;
+
+  const Nullable(this._value);
+
+  T? get value => _value;
+
+  bool get isNotNull => _value != null;
+}