+++ dio_cache_interceptor/test/cache_interceptor_test.dart
@@ -1,5 +1,6 @@
 import 'package:dio/dio.dart';
 import 'package:dio_cache_interceptor/dio_cache_interceptor.dart';
+import 'package:dio_cache_interceptor/src/model/nullable.dart';
 import 'package:test/test.dart';
 
 import 'mock_httpclient_adapter.dart';
@@ -21,60 +22,6 @@ void main() {
 
   tearDown(() async {
     _dio.close();
-  });
-
-  test('maxStale', () async {
-    final resp = await _dio.get(
-      '${MockHttpClientAdapter.mockBase}/ok-nodirective',
-      options: options
-          .copyWith(
-              policy: CachePolicy.forceCache,
-              maxStale: const Duration(seconds: 1))
-          .toOptions(),
-    );
-    expect(resp.statusCode, equals(200));
-    expect(resp.extra[CacheResponse.cacheKey], isNotNull);
-  });
-
-  test('maxStale removal', () async {
-    // Request for the 1st time
-    var resp = await _dio.get(
-      '${MockHttpClientAdapter.mockBase}/ok-nodirective',
-      options: options
-          .copyWith(
-            policy: CachePolicy.forceCache,
-            maxStale: const Duration(seconds: 1),
-          )
-          .toOptions(),
-    );
-    var key = resp.extra[CacheResponse.cacheKey];
-    expect(await store.exists(key), isTrue);
-
-    await Future.delayed(const Duration(seconds: 2));
-
-    // Request a 2nd time to postpone stale date
-    // We wait for 2 second so the cache is now staled but we recover it
-    resp = await _dio.get(
-      '${MockHttpClientAdapter.mockBase}/ok-nodirective',
-      options: options
-          .copyWith(
-            policy: CachePolicy.forceCache,
-            maxStale: const Duration(seconds: 1),
-          )
-          .toOptions(),
-    );
-    expect(await store.exists(key), isTrue);
-
-    await Future.delayed(const Duration(seconds: 1));
-
-    // Request for the last time without maxStale directive to ensure
-    // the cache entry is now deleted
-    resp = await _dio.get(
-      '${MockHttpClientAdapter.mockBase}/ok-nodirective',
-      options: options.toOptions(),
-    );
-
-    expect(await store.exists(key), isFalse);
   });
 
   test('Fetch stream 200', () async {
@@ -105,12 +52,12 @@ void main() {
 
   test('Fetch with cipher', () async {
     final cipherOptions = options.copyWith(
-      cipher: CacheCipher(
+      cipher: Nullable(CacheCipher(
         decrypt: (bytes) =>
             Future.value(bytes.reversed.toList(growable: false)),
         encrypt: (bytes) =>
             Future.value(bytes.reversed.toList(growable: false)),
-      ),
+      )),
     );
 
     var resp = await _dio.get(
@@ -198,7 +145,7 @@ void main() {
       options: options
           .copyWith(
             policy: CachePolicy.refresh,
-            maxStale: Duration(minutes: 10),
+            maxStale: Nullable(Duration(minutes: 10)),
           )
           .toOptions(),
     );
@@ -235,9 +182,11 @@ void main() {
       await _dio.get(
         '${MockHttpClientAdapter.mockBase}/ok',
         options: Options(
-          extra: options.copyWith(
-              hitCacheOnErrorExcept: [500],
-              policy: CachePolicy.refresh).toExtra()
+          extra: options
+              .copyWith(
+                  hitCacheOnErrorExcept: Nullable([500]),
+                  policy: CachePolicy.refresh)
+              .toExtra()
             ..addAll({'x-err': '500'}),
         ),
       );
@@ -274,10 +223,12 @@ void main() {
     final resp2 = await _dio.get(
       '${MockHttpClientAdapter.mockBase}/ok',
       options: Options(
-        extra: options.copyWith(
-          hitCacheOnErrorExcept: [],
-          policy: CachePolicy.refresh,
-        ).toExtra()
+        extra: options
+            .copyWith(
+              hitCacheOnErrorExcept: Nullable([]),
+              policy: CachePolicy.refresh,
+            )
+            .toExtra()
           ..addAll({'x-err': '500'}),
       ),
     );
@@ -294,9 +245,7 @@ void main() {
     final resp2 = await _dio.get(
       '${MockHttpClientAdapter.mockBase}/exception',
       options: Options(
-        extra: options.copyWith(
-          hitCacheOnErrorExcept: [],
-        ).toExtra()
+        extra: options.copyWith(hitCacheOnErrorExcept: Nullable([])).toExtra()
           ..addAll({'x-err': '500'}),
       ),
     );